智知因/requirements.txt
# 智知因 - 基于大模型的多智能体个性化学习系统

fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0

# LLM框架
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-core>=0.1.0

# 向量数据库
chromadb>=0.4.18
sentence-transformers>=2.2.2

# 数据库
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# 文档处理
pypdf>=3.17.0
python-docx>=1.1.0
python-multipart>=0.0.6

# 配置管理
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# 日志
loguru>=0.7.0

# 其他工具
httpx>=0.25.0
aiofiles>=23.2.0
